Refactor schema tester plugin to avoid circular imports
